Migrate tests to python 3.11
Use unittest from standart lib as main lib for testing
Move .py file with fixtures to subdir to avoid relative import (tests)
Update requierements: use newer versions of numpy and pandas
Fix bug with depreciated module collections.Container
