+++ requirements.txt
@@ -1,10 +1,9 @@
-pandas==1.3.1
+numpy==1.24.2
+pandas==1.4.2
 requests==2.24.0
 click==7.1.2
 click-datetime==0.2
 # dev
-nose
-mock
 coverage
 urltools
 parameterized
